+++ server/entities/ammo_buckshot.cpp
@@ -1,0 +1,39 @@
+/***
+*
+*	Copyright (c) 1996-2002, Valve LLC. All rights reserved.
+*
+*	This product contains software technology licensed from Id
+*	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc.
+*	All Rights Reserved.
+*
+*   Use, distribution, and modification of this source code and/or resulting
+*   object code is restricted to non-commercial enhancements to products from
+*   Valve LLC.  All other use, distribution, or modification is prohibited
+*   without written permission from Valve LLC.
+*
+****/
+
+#include "ammo_buckshot.h"
+
+LINK_ENTITY_TO_CLASS(ammo_buckshot, CShotgunAmmo);
+
+void CShotgunAmmo::Spawn(void)
+{
+	Precache();
+	SET_MODEL(ENT(pev), "models/w_shotbox.mdl");
+	BaseClass::Spawn();
+}
+void CShotgunAmmo::Precache(void)
+{
+	PRECACHE_MODEL("models/w_shotbox.mdl");
+	PRECACHE_SOUND("items/9mmclip1.wav");
+}
+BOOL CShotgunAmmo::AddAmmo(CBaseEntity *pOther)
+{
+	if (pOther->GiveAmmo(AMMO_BUCKSHOTBOX_GIVE, "buckshot", BUCKSHOT_MAX_CARRY) != -1)
+	{
+		EMIT_SOUND(ENT(pev), CHAN_ITEM, "items/9mmclip1.wav", 1, ATTN_NORM);
+		return TRUE;
+	}
+	return FALSE;
+}
